batch and select get their own files here too
